+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 [Full Changelog] [Previous Releases]
 
+## [1.1.1] - 2023-05-21
+
+### Added
+- Set option to silence TheGrimRepair Utilities window messages as default
+
+### Fixed
+- Set TheGrimRepair Utilities to NOT show with merchant window by default (users can still change this option)
+
 ## [1.1.0] - 2023-05-21
 
 ### Added
@@ -111,7 +119,7 @@
 - Auto repairs using personal funds or guild repairs (if available)
 - Auto sells gray items, optionally keep items which should become eligible for transmog in a future WoW update
 
-[Full Changelog]: https://github.com/Br3gGames/TheGrimRepair/compare/1.0.2...1.1.0
+[Full Changelog]: https://github.com/Br3gGames/TheGrimRepair/compare/1.0.2...1.1.1
 [Previous Releases]: https://github.com/Br3gGames/TheGrimRepair/releases
 [0.0.1]: https://github.com/Br3gGames/TheGrimRepair/releases/tag/0.0.1
 [0.0.2]: https://github.com/Br3gGames/TheGrimRepair/releases/tag/0.0.2
@@ -128,3 +136,4 @@
 [1.0.1]: https://github.com/Br3gGames/TheGrimRepair/releases/tag/1.0.1
 [1.0.2]: https://github.com/Br3gGames/TheGrimRepair/releases/tag/1.0.2
 [1.1.0]: https://github.com/Br3gGames/TheGrimRepair/releases/tag/1.1.0
+[1.1.1]: https://github.com/Br3gGames/TheGrimRepair/releases/tag/1.1.1